ci: Split base.py into 3 files
