+++ tests/integration/unit_test/test_unit_test_python3_8.py
@@ -1,17 +1,17 @@
-from tests.integration.base import Base, LOG
+from tests.integration.unit_test.unit_test_base import UnitTestBase
 
 
-class UnitTest_python3_8_cookiecutter_aws_sam_hello_python(Base.Python38UnitTestBase):
+class UnitTest_python3_8_cookiecutter_aws_sam_hello_python(UnitTestBase.Python38UnitTestBase):
     directory = "python3.8/cookiecutter-aws-sam-hello-python"
     code_directories = ["hello_world"]
 
 
-class UnitTest_python3_8_cookiecutter_aws_sam_eventBridge_python(Base.Python38UnitTestBase):
+class UnitTest_python3_8_cookiecutter_aws_sam_eventBridge_python(UnitTestBase.Python38UnitTestBase):
     directory = "python3.8/cookiecutter-aws-sam-eventBridge-python"
     code_directories = ["hello_world_function"]
 
 
-class UnitTest_python3_8_cookiecutter_aws_sam_eventbridge_schema_app_python(Base.Python38UnitTestBase):
+class UnitTest_python3_8_cookiecutter_aws_sam_eventbridge_schema_app_python(UnitTestBase.Python38UnitTestBase):
     directory = "python3.8/cookiecutter-aws-sam-eventbridge-schema-app-python"
     code_directories = ["hello_world_function"]
 
@@ -20,7 +20,7 @@ class UnitTest_python3_8_cookiecutter_aws_sam_eventbridge_schema_app_python(Base
         pass
 
 
-class UnitTest_python3_8_cookiecutter_aws_sam_step_functions_sample_app(Base.Python38UnitTestBase):
+class UnitTest_python3_8_cookiecutter_aws_sam_step_functions_sample_app(UnitTestBase.Python38UnitTestBase):
     directory = "python3.8/cookiecutter-aws-sam-step-functions-sample-app"
     code_directories = [
         "functions/stock_buyer",
@@ -29,6 +29,6 @@ class UnitTest_python3_8_cookiecutter_aws_sam_step_functions_sample_app(Base.Pyt
     ]
 
 
-class UnitTest_python3_8_cookiecutter_aws_sam_efs_python(Base.Python38UnitTestBase):
+class UnitTest_python3_8_cookiecutter_aws_sam_efs_python(UnitTestBase.Python38UnitTestBase):
     directory = "python3.8/cookiecutter-aws-sam-efs-python"
     code_directories = ["hello_efs"]
